Database class added
